Bug fixed
