--- src/pages/Home/Company/companiesLogosData.js
@@ -1,8 +1,0 @@
-// All companies logos
-export const logos = [
-      { id: 1, src: "/asset/img/google.png", alt: "Google" },
-      { id: 2, src: "/asset/img/meta.png", alt: "Meta" },
-      { id: 3, src: "/asset/img/netflix.png", alt: "Netflix" },
-      { id: 4, src: "/asset/img/microsoft2.png", alt: "Microsoft" },
-      { id: 5, src: "/asset/img/paypal.png", alt: "PayPal" },
-    ];
