+++ src/pages/Home/KeyFeature/KeyFeatures.js
@@ -1,15 +1,33 @@
 "use client";
 import SectionTitle from "@/components/Common/sectionTitle/SectionTitle";
-import {
-  collaborationArray,
-  customizationArray,
-  taskManagementArray,
-} from "./keyFeatureData";
+// import taskManagementArray, { collaborationArray, customizationArray } from "./keyFeatureData";
 // import style from "./key.module.css"
 import ArrowBoldSign from "./ArrowBoldSign";
 import { IoMdCheckmarkCircle } from "react-icons/io";
 import { IoMdCheckmarkCircleOutline } from "react-icons/io";
 import GradientBg from "@/components/Common/gradientBg/GradientBg";
+
+
+const taskManagementArray = [
+  { id: 1, text: "Intuitive Task Creation and Management" },
+  { id: 2, text: "Experience Effortless Task Management with our Intuitive Interface." },
+  { id: 3, text: "Seamlessly Organize Your Tasks for Enhanced Productivity" },
+  { id: 4, text: "Unlock Efficiency with our Streamlined Task Creation Process" }
+];
+
+const collaborationArray = [
+  { id: 5, text: "Real-time Collaboration and Communication" },
+  { id: 6, text: "Enhance Team Collaboration with Instant Communication Features." },
+  { id: 7, text: "Create an Instant Communication Hub for Seamless Teamwork." },
+  { id: 8, text: "Achieve Real-Time Collaboration for Increased Project Efficiency." }
+];
+
+export const customizationArray = [
+  { id: 9, text: "Customizable Workflows and Boards." },
+  { id: 10, text: "Tailor Workflows to Fit Your Project's Unique Requirements." },
+  { id: 11, text: "Flexible Project Management Boards for Your Changing Needs." },
+  { id: 12, text: "Create Your Ideal Workflow with Customizable Features." }
+];
 
 
 const KeyFeatures = () => {
