--- src/pages/Home/KeyFeature/keyFeatureData.js
@@ -1,21 +1,0 @@
-export const taskManagementArray = [
-      { id: 1, text: "Intuitive Task Creation and Management" },
-      { id: 2, text: "Experience Effortless Task Management with our Intuitive Interface." },
-      { id: 3, text: "Seamlessly Organize Your Tasks for Enhanced Productivity" },
-      { id: 4, text: "Unlock Efficiency with our Streamlined Task Creation Process" }
-    ];
-    
-    export const collaborationArray = [
-      { id: 5, text: "Real-time Collaboration and Communication" },
-      { id: 6, text: "Enhance Team Collaboration with Instant Communication Features." },
-      { id: 7, text: "Create an Instant Communication Hub for Seamless Teamwork." },
-      { id: 8, text: "Achieve Real-Time Collaboration for Increased Project Efficiency." }
-    ];
-    
-    export const customizationArray = [
-      { id: 9, text: "Customizable Workflows and Boards." },
-      { id: 10, text: "Tailor Workflows to Fit Your Project's Unique Requirements." },
-      { id: 11, text: "Flexible Project Management Boards for Your Changing Needs." },
-      { id: 12, text: "Create Your Ideal Workflow with Customizable Features." }
-    ];
-
